+++ src/main/kotlin/ar/edu/unq/agiletutor/service/StudentService.kt
@@ -153,7 +153,7 @@ class StudentService {
     }
 
     @Transactional
-    fun studentsNotBlockedAbsentAtAParticularDay(courseId: Int, day: Int): List<Student> {
+    fun studentsNotBlockedAbsentAtAParticularDay(courseId: Long, day: Int): List<Student> {
         val rangedays = (1..6)
         if (!rangedays.contains(day)) {
             throw ItemNotFoundException(" Day:  $day invalid")
